+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
-
+- Add `capacityType` and `taints` to `ManagedNodeGroup`
+  [#614](https://github.com/pulumi/pulumi-eks/pull/614)
 
 ## 0.32.0 (Released August 4, 2021)